Fix UAF while adding items to a context.
When xlog_cil_ail_insert() loops to create all the batches, the context
could have been emptied before moving to the new batch. This causes the
context to be freed by the last item being removed.

This patch ensures we pin the context in memory until we are done adding
all the items before attempting to free it.

This is not an individual patch, but will probably be squashed into
something else, it's more to register this detail.

Also, I'm not sure if this could implicate on a context never being
freed *if* we just have a single batch.

So far it passed all generic tests until 392, which tests recovery, but
I was already expecting that.