xfs_log_item: Fix a couple items mishandling
generic/392 was crashing on an invalid reference to the item li_ops.
Initially it led me to believe this was a UAF, but in fact the problem
was relatedto a couple bugs in the AIL traversal helpers:

xfs_trans_ail_cursor_first() was still wired to deal with a singly
linked list of AIL items. Fixing this was not complicated, as we only
need to traverse both ail list and each context fo look for the first
LSN greater than the one passed.

The biggest issue was in xfs_ail_next(). We need to be careful on this
function, as the item we've got, might be the last one in the current
context, so we need to continue the search on the following contexts.

This was originally returning the first entry of the next context, but
was not taking into account the next context might be null. It ended up
returning a pointer to the list head instead of the item per se.

Signed-off-by: Carlos Maiolino <cem@kernel.org>