Fix a spurious context in the AIL
If we create a new context but end up not adding any item to it, there
will be no way for the item to be removed from the ail, as it is
currently being freed once its last item has been deleted.
If no items exist in the AIL, no items will be deleted.

Fix this by not adding the context to the ail if it doesn't have any
item within it (we shouldn't be adding an empty context to the ail
anyway), and free it.

Signed-off-by: Carlos Maiolino <cmaiolino@redhat.com>