Prevent context to be freed during recovery
This is a hack to avoid a context to be freed while items are being
added to the AIL.

The issue is that the same transaction can have items that will be added
and subsequentially deleted from the ail context, causing the AIL
context to be freed before all items are processed.

I'm still not sure what's the best way to handle it, so, just hack it
until I come up with a better idea.